날짜 중간 저장

+++ src/main/java/com/example/trello/service/TaskService.java
@@ -4,8 +4,10 @@ import com.example.trello.dto.RestApiResponseDto;
 import com.example.trello.dto.TaskRequestDto;
 import com.example.trello.dto.TaskResponseDto;
 import com.example.trello.entity.Card;
+import com.example.trello.entity.Columns;
 import com.example.trello.entity.Task;
 import com.example.trello.repository.CardRepository;
+import com.example.trello.repository.ColumnsRepository;
 import com.example.trello.repository.TaskRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,8 +29,21 @@ public class TaskService {
     @Autowired
     private CardRepository cardRepository;
 
-    public ResponseEntity<RestApiResponseDto> getTask(Long cardId) {
+    @Autowired
+    private ColumnsRepository columnsRepository;
+
+    public ResponseEntity<RestApiResponseDto> getTask(Long columnId, Long cardId) {
         try {
+            Columns columns = columnsRepository.findById(columnId)
+                    .orElseThrow(() -> new IllegalArgumentException("해당 컬럼이 존재하지 않습니다."));
+
+            Card card = cardRepository.findById(cardId).orElseThrow(
+                    () -> new IllegalArgumentException("해당 카드가 존재하지않습니다."));
+
+            if (!card.getColumns().getId().equals(columnId) || !card.getColumns().getId().equals(columns.getId())) {
+                throw new IllegalArgumentException("해당 카드는 지정된 컬럼에 속해 있지 않습니다.");
+            }
+
             Task task = taskRepository.findByCardId(cardId).orElseThrow(
                     () -> new IllegalArgumentException("해당 카드에 설정된 날짜가 없습니다."));
 
@@ -40,31 +55,51 @@ public class TaskService {
         }
     }
 
-    public ResponseEntity<RestApiResponseDto> createCardDates(Long cardId, TaskRequestDto requestDto) {
+    public ResponseEntity<RestApiResponseDto> createCardDates(Long columnId, Long cardId, TaskRequestDto requestDto) {
         try {
+            Columns columns = columnsRepository.findById(columnId)
+                    .orElseThrow(() -> new IllegalArgumentException("해당 컬럼이 존재하지 않습니다."));
+
             Card card = cardRepository.findById(cardId).orElseThrow(
                     () -> new IllegalArgumentException("해당 카드가 존재하지않습니다."));
 
-            Task task = new Task(requestDto, card);
-
-            Optional<Task> chackTask = taskRepository.findByCardId(cardId);
-
-            if (chackTask.isPresent()) {
-                log.info("중복 테스트");
-                throw new IllegalArgumentException("해당 카드에 이미 설정된 날짜가 있습니다.");
+            if (!card.getColumns().getId().equals(columnId) || !card.getColumns().getId().equals(columns.getId())) {
+                throw new IllegalArgumentException("해당 카드는 지정된 컬럼에 속해 있지 않습니다.");
             }
+
+            Optional<Task> chackColumnTask = taskRepository.findByCardId(columnId);
+
+            Optional<Task> chackCardTask = taskRepository.findByCardId(cardId);
+
+            if (chackColumnTask.isPresent() && chackCardTask.isPresent()) {
+                log.info("중복 테스트");
+                throw new IllegalArgumentException("해당 컬럼안 카드에 이미 설정된 날짜가 있습니다.");
+            }
+
+            Task task = new Task(requestDto, card, columns);
+
             task.setStartDate(LocalDateTime.now());
 
             taskRepository.save(task);
 
-            return this.resultResponse(HttpStatus.CREATED, " 날짜 설정 완료", new TaskResponseDto(task));
+            return this.resultResponse(HttpStatus.CREATED, "날짜 설정 완료", new TaskResponseDto(task));
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body(new RestApiResponseDto(HttpStatus.BAD_REQUEST.value(), e.getMessage()));
         }
     }
 
-    public ResponseEntity<RestApiResponseDto> updateCardDates(Long dateId, TaskRequestDto requestDto) {
+    public ResponseEntity<RestApiResponseDto> updateCardDates(Long columnId, Long cardId, Long dateId, TaskRequestDto requestDto) {
         try {
+            Columns columns = columnsRepository.findById(columnId)
+                    .orElseThrow(() -> new IllegalArgumentException("해당 컬럼이 존재하지 않습니다."));
+
+            Card card = cardRepository.findById(cardId).orElseThrow(
+                    () -> new IllegalArgumentException("해당 카드가 존재하지않습니다."));
+
+            if (!card.getColumns().getId().equals(columnId) || !card.getColumns().getId().equals(columns.getId())) {
+                throw new IllegalArgumentException("해당 카드는 지정된 컬럼에 속해 있지 않습니다.");
+            }
+
             Task task = taskRepository.findById(dateId).orElseThrow(
                     () -> new IllegalArgumentException("설정한 날짜가 존재하지 않습니다."));
 
@@ -73,20 +108,30 @@ public class TaskService {
 
             task = taskRepository.save(task);
 
-            return this.resultResponse(HttpStatus.CREATED, " 날짜 수정 완료", new TaskResponseDto(task));
+            return this.resultResponse(HttpStatus.OK, "날짜 수정 완료", new TaskResponseDto(task));
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body(new RestApiResponseDto(HttpStatus.BAD_REQUEST.value(), e.getMessage()));
         }
     }
 
-    public ResponseEntity<RestApiResponseDto> deleteCardDates(Long dateId) {
+    public ResponseEntity<RestApiResponseDto> deleteCardDates(Long columnId, Long cardId, Long dateId) {
         try {
+            Columns columns = columnsRepository.findById(columnId)
+                    .orElseThrow(() -> new IllegalArgumentException("해당 컬럼이 존재하지 않습니다."));
+
+            Card card = cardRepository.findById(cardId).orElseThrow(
+                    () -> new IllegalArgumentException("해당 카드가 존재하지않습니다."));
+
+            if (!card.getColumns().getId().equals(columnId) || !card.getColumns().getId().equals(columns.getId())) {
+                throw new IllegalArgumentException("해당 카드는 지정된 컬럼에 속해 있지 않습니다.");
+            }
+
             Task task = taskRepository.findById(dateId).orElseThrow(
                     () -> new IllegalArgumentException("설정한 날짜가 존재하지 않습니다."));
 
             taskRepository.delete(task);
 
-            return this.resultResponse(HttpStatus.CREATED, " 날짜 초기화 완료", null);
+            return this.resultResponse(HttpStatus.OK, "날짜 초기화 완료", null);
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body(new RestApiResponseDto(HttpStatus.BAD_REQUEST.value(), e.getMessage()));
         }
